+++ src/Cvoya.Spring.Host.Worker/Program.cs
@@ -5,6 +5,7 @@ using System.Runtime.InteropServices;
 
 using Cvoya.Spring.Connector.Arxiv.DependencyInjection;
 using Cvoya.Spring.Connector.GitHub.DependencyInjection;
+using Cvoya.Spring.Connector.WebSearch.DependencyInjection;
 using Cvoya.Spring.Dapr.Actors;
 using Cvoya.Spring.Dapr.DependencyInjection;
 using Cvoya.Spring.Dapr.Workflows;
@@ -51,7 +52,8 @@ builder.Services
     .AddCvoyaSpringDapr(builder.Configuration)
     .AddCvoyaSpringOllamaLlm(builder.Configuration)
     .AddCvoyaSpringConnectorGitHub(builder.Configuration)
-    .AddCvoyaSpringConnectorArxiv(builder.Configuration);
+    .AddCvoyaSpringConnectorArxiv(builder.Configuration)
+    .AddCvoyaSpringConnectorWebSearch(builder.Configuration);
 
 // DataProtection registration is gated by design-time tooling to avoid
 // noisy ephemeral-key warnings during build-time OpenAPI generation. The